requirements.txt
dgl-cu110==0.6.1
torch==1.10.0
rich
pandas
numpy
cppimport
tqdm
transformers
matplotlib
networkx
tensorboardX
sentencepiece
scikit-learn